Separation of jvm-specific code from TableQuerySelectConnectionTest
